move loading into own file
